\name{copulaIVTS-package}
\alias{copulaIVTS-package}
\alias{copulaIVTS}
\docType{package}
\title{
Gaussian copula model for Integer-valued Time series - Maximum Likelihood estimation using the MCEM
}
\description{
The R package copulaIVTS provides all statistical R used to implement a Monte Carlo Expectation Maximisation algorithm for maximum likelihood estimation of the dependence parameters of a digitised Gaussian ARMA model. Useful for count time series also.

Monte Carlo EM algorithm Code.
Code to run the Monte Carlo EM algorithm for parameter estimation of the Gaussian copula model with specified discrete marginals.

To be be added to the R package shortly are two alternative estimation methods to fit the digitised Gaussian ARMA model provided in the Thesis link
via Approximate Bayesian Computation (ABC) and
via approximate MLE via a d-vine bivariate copula representation.
}
\details{
\tabular{ll}{
Package: \tab copulaIVTS\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2021-03-23\cr
License: \tab GPL (>=2) \cr
}
}
\author{
Hannah Lennon
<drhannahlennon@gmail.com>
}
\references{
Lennon H., & Yuan J., Estimation of a digitised Gaussian ARMA model by Monte Carlo Expectation Maximisation, Computational Statistics & Data Analysis 2018;8:e020683

Lennon, H., 2016. Gaussian copula modelling for integer-valued time series (Doctoral Thesis, University of Manchester).
}
\keyword{ package }
\seealso{
}
\examples{
}
