\name{MCEMalg-package}
\alias{MCEMalg-package}
\alias{MCEMalg}
\docType{package}
\title{
Maximum Likelihood estimation using the MCEM
}
\description{
Monte Carlo EM algorithm Code.
Code to run the Monte Carlo EM algorithm for parameter estimation of the Gaussian copula model with specified discrete marginals.
}
\details{
\tabular{ll}{
Package: \tab MCEMalg\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-05-04\cr
License: \tab GPL (>=2) \cr
}
}
\author{
Hannah Lennon
<drhannahlennon@gmail.com>
}
\references{
Lennon H., & Yuan J., Estimation of a digitised Gaussian ARMA model by Monte Carlo Expectation Maximisation, Computational Statistics & Data Analysis 2018;8:e020683

Lennon, H., 2016. Gaussian copula modelling for integer-valued time series (Doctoral Thesis, University of Manchester).
}
\keyword{ package }
\seealso{
}
\examples{
a <- 9 + 3}
